+++ src/api/consumer/lib/inject-conf.js
@@ -1,0 +1,15 @@
+/** @flow */
+
+import { loadConsumer, Consumer } from '../../../consumer';
+import GeneralError from '../../../error/general-error';
+import type { InjectConfResult } from '../../../consumer/component-ops/inject-conf';
+
+export default (async function injectConf(id: string, force: boolean): Promise<InjectConfResult> {
+  if (!id) {
+    throw new GeneralError('please specify component id');
+  }
+  const consumer: Consumer = await loadConsumer();
+  const attachResults = await consumer.injectConf(consumer.getParsedId(id), force);
+  await consumer.onDestroy();
+  return attachResults;
+});