+++ src/consumer/versions-ops/checkout-version.js
@@ -203,7 +203,7 @@ async function applyVersion(
     componentsWithDependencies,
     installNpmPackages: shouldInstallNpmPackages(),
     override: true,
-    writeBitJson: !!componentFromFS.bitJson || componentMap.configDir, // write bit.json and config files only if it was there before
+    writeConfig: !!componentFromFS.bitJson || componentMap.configDir, // write bit.json and config files only if it was there before
     configDir: componentMap.configDir,
     verbose,
     writeDists: !ignoreDist,